+++ src/modules/home/index.ts
@@ -1,0 +1,1 @@
+export { default as HomeBody } from "./views/HomeBody";
